requirements.txt
streamlit>=1.24
numpy
pillow
opencv-contrib-python-headless
google-cloud-vision
python-dotenv
pandas
plotly
streamlit
google-api-core
networkx

